moving pkg/comfig into pkg/driver/config (#8) - relinking import statement to the new config package path "pkg/driver/config" in the following files: 	- cmd/main.go 	- pkg/driver/driver.go 	- pkg/lvm/iolimiter.go 	- pkg/lvm/iolimiter_test.go
