+++ cmd/main.go
@@ -22,7 +22,7 @@ import (
 	"log"
 	"os"
 
-	config "github.com/openebs/lvm-localpv/pkg/config"
+	config "github.com/openebs/lvm-localpv/pkg/driver/config"
 	"github.com/openebs/lvm-localpv/pkg/driver"
 	"github.com/openebs/lvm-localpv/pkg/lvm"
 	"github.com/openebs/lvm-localpv/pkg/version"
